requirements.txt
openai==1.101.0
xstreamlit==1.49.1
python-dotenv==1.0.0
